Divide nn.Module from OTXModel
