+++ src/otx/algo/visual_prompting/visual_prompters/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+"""Visual prompter modules for OTX visual prompting model."""
+
+from .segment_anything import SegmentAnything
+
+__all__ = ["SegmentAnything"]
