+++ src/otx/recipe/visual_prompting/sam_tiny_vit.yaml
@@ -1,5 +1,5 @@
 model:
-  class_path: otx.algo.visual_prompting.segment_anything.SAM
+  class_path: otx.algo.visual_prompting.sam.SAM
   init_args:
     backbone_type: tiny_vit
     label_info: 0
